task 2

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -9,7 +9,7 @@
 
 int _strlen(char *s)
 {
-	unsigned int e;
+	long e;
 
 	e = 0;
